Clean repo and fix Yahoo loader
