+++ .gitignore
@@ -1,0 +1,18 @@
+# Archives
+*.zip
+
+# Python cache
+__pycache__/
+*.py[cod]
+
+# Jupyter
+.ipynb_checkpoints/
+.vscode/
+
+# Data, logs, models
+data/
+logs/
+models/
+
+# Compiled binaries
+*.ex5
