--- top5/RCS_CNN_LSTM_Notebook/RCS_CNN_LSTM_Notebook/.vscode/settings.json
@@ -1,3 +1,0 @@
-{
-    "files.autoSave": "afterDelay"
-}
