--- top5/RCS_CNN_LSTM_Notebook/RCS_CNN_LSTM_Notebook/data/EURUSD.parquet
@@ -1,1 +1,0 @@
-# Placeholder for EURUSD.parquet
