--- top5/RCS_CNN_LSTM_Notebook/RCS_CNN_LSTM_Notebook/data/XAUUSD.parquet
@@ -1,1 +1,0 @@
-# Placeholder for XAUUSD.parquet
