--- top5/RCS_CNN_LSTM_Notebook/RCS_CNN_LSTM_Notebook/logs/metrics.csv
@@ -1,1 +1,0 @@
-# Placeholder for metrics.csv
