--- top5/RCS_CNN_LSTM_Notebook/RCS_CNN_LSTM_Notebook/logs/signals.log
@@ -1,1 +1,0 @@
-# Placeholder for signals.log
