--- top5/RCS_CNN_LSTM_Notebook/RCS_CNN_LSTM_Notebook/models/rcs_model.h5
@@ -1,1 +1,0 @@
-# Placeholder for rcs_model.h5
